uv/tests: TEMPORARY
This is correct insomuch as the current implementation does
not support groups. Since the current impl doesn't see groups,
and doesn't know about them in "world knowledge," it thinks
it needs to specify `extra != 'project2'` when the `project1`
group is enabled. But this is superfluous because `project1`
and `project2` are declared as conflicting.